refactor: segregate gstr-1 common utility
